TT method works

+++ src/timetagger.h
@@ -42,7 +42,7 @@ void TTsetTriggerLevel(TimeTaggerNetwork *t, int32_t channel, double level);
 std::unique_ptr<Correlation> TTcreateCorrelation(TimeTaggerNetwork *t, int32_t channel1, int32_t channel2, int32_t bin_width, int32_t max_count);
 std::unique_ptr<std::vector<int32_t>> CorrelationGetData(Correlation &c);
 
-std::unique_ptr<Counter> TTcreateCounter(TimeTaggerNetwork *t, rust::Vec<int32_t> channels, double bin_width, int32_t max_count);
+std::unique_ptr<Counter> TTcreateCounter(TimeTaggerNetwork *t, rust::Vec<int32_t> channels, int32_t bin_width, int32_t max_count);
 std::unique_ptr<std::vector<int32_t>> CounterGetData(Counter &c);
 void CorrelationStart(Correlation &c);
 void CorrelationStartFor(Correlation &c, int64_t capture_duration, bool clear);
